refactor(id): standardize of entity id to long

+++ src/main/java/com/ionidea/mothramxbe/tasks/repository/LeaveEntryRepository.java
@@ -3,6 +3,6 @@ package com.ionidea.mothramxbe.tasks.repository;
 import com.ionidea.mothramxbe.tasks.model.LeaveEntry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaveEntryRepository extends JpaRepository<LeaveEntry, Integer> {
+public interface LeaveEntryRepository extends JpaRepository<LeaveEntry, Long> {
 
 }
