+++ src/main/java/com/ionidea/mothramxbe/tasks/repository/LeaveTypeRepository.java
@@ -3,6 +3,6 @@ package com.ionidea.mothramxbe.tasks.repository;
 import com.ionidea.mothramxbe.tasks.model.LeaveType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaveTypeRepository extends JpaRepository<LeaveType, Integer> {
+public interface LeaveTypeRepository extends JpaRepository<LeaveType, Long> {
 
 }
