+++ src/main/java/com/ionidea/mothramxbe/tasks/repository/TaskRefMonthRepository.java
@@ -3,6 +3,6 @@ package com.ionidea.mothramxbe.tasks.repository;
 import com.ionidea.mothramxbe.system.entity.RefMonth;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskRefMonthRepository extends JpaRepository<RefMonth, Integer> {
+public interface TaskRefMonthRepository extends JpaRepository<RefMonth, Long> {
 
 }
